Refactor : main.basic.js에서 dom요소와 핸들러 분리
